move keadm init/join/reset to deprecated
