+++ .github/workflows/main.yaml
@@ -177,10 +177,10 @@ jobs:
             /tmp/cloudcore.log
             /tmp/edgecore.log
 
-  keadm_e2e_test:
+  keadm_deprecated_e2e_test:
     runs-on: ubuntu-18.04
     timeout-minutes: 30
-    name: Keadm e2e test
+    name: Keadm deprecated e2e test
     needs: image-prepare
     env:
       GO111MODULE: on
@@ -216,7 +216,7 @@ jobs:
         run: |
           docker load < /home/runner/build-tools/build-tools.tar
 
-      - run: make keadm_e2e
+      - run: make keadm_deprecated_e2e
 
   keadm_beta_e2e_test:
     runs-on: ubuntu-18.04
